Update Secret Manager samples to be canonical
