+++ secretmanager/pom.xml
@@ -17,7 +17,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.google.cloud.secretmanager.samples</groupId>
+  <groupId>secretmanager</groupId>
   <artifactId>secretmanager-samples</artifactId>
   <packaging>jar</packaging>
