+++ secretmanager/src/main/java/secretmanager/IamGrantAccess.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example;
+package secretmanager;
 
 // [START secretmanager_iam_grant_access]
 import com.google.cloud.secretmanager.v1.SecretManagerServiceClient;
@@ -42,14 +42,12 @@ public class IamGrantAccess {
     // the "close" method on the client to safely clean up any remaining background resources.
     try (SecretManagerServiceClient client = SecretManagerServiceClient.create()) {
       // Build the name from the version.
-      SecretName name = SecretName.of(projectId, secretId);
-
-      // Create a request to get the current IAM policy.
-      GetIamPolicyRequest getRequest =
-          GetIamPolicyRequest.newBuilder().setResource(name.toString()).build();
+      SecretName secretName = SecretName.of(projectId, secretId);
 
       // Request the current IAM policy.
-      Policy currentPolicy = client.getIamPolicy(getRequest);
+      Policy currentPolicy =
+          client.getIamPolicy(
+              GetIamPolicyRequest.newBuilder().setResource(secretName.toString()).build());
 
       // Build the new binding.
       Binding binding =
@@ -61,15 +59,12 @@ public class IamGrantAccess {
       // Create a new IAM policy from the current policy, adding the binding.
       Policy newPolicy = Policy.newBuilder().mergeFrom(currentPolicy).addBindings(binding).build();
 
-      // Create a request to update the IAM policy.
-      SetIamPolicyRequest setRequest =
-          SetIamPolicyRequest.newBuilder()
-              .setResource(name.toString())
-              .setPolicy(newPolicy)
-              .build();
-
       // Save the updated IAM policy.
-      client.setIamPolicy(setRequest);
+      client.setIamPolicy(
+          SetIamPolicyRequest.newBuilder()
+              .setResource(secretName.toString())
+              .setPolicy(newPolicy)
+              .build());
 
       System.out.printf("Updated IAM policy for %s\n", secretId);
     }
