+++ secretmanager/src/test/java/secretmanager/SnippetsIT.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example;
+package secretmanager;
 
 import static com.google.common.truth.Truth.assertThat;
